+++ game_patch/misc/alpine_settings.h
@@ -91,7 +91,8 @@ struct AlpineGameSettings
     bool try_disable_muzzle_flash_lights = false;
     bool world_hud_ctf_icons = true;
     bool world_hud_alt_damage_indicators = false;
-    bool world_hud_overdraw = true;
+    bool world_hud_flag_overdraw = true;
+    bool world_hud_hill_overdraw = true;
     bool world_hud_damage_numbers = true;
     bool world_hud_spectate_player_labels = false;
     bool world_hud_team_player_labels = false;